feat(swipe-and-profile): add a swipe feature to the home screen and a profile details screen
This commit adds additional SETUP to the project such as prettier, commitzen and typescript aliases.
It also introduces the swipe left/right feature on the home screen and the option to tap and view
the user's profile. New native dependencies have been added to the project which turns necessary a
new binary in order to release.

BREAKING CHANGE: Added new native dependencies to the project: react-native-reanimated,
react-native-gesture-handler, react-native-linear-gradient, react-native-fast-image,
react-native-shared-element